backend/requirements.txt
Flask==3.0.3
Flask-Cors==3.0.10
psycopg2-binary==2.9.9
python-dotenv==1.0.0
importlib-metadata==8.1.0
Werkzeug==3.0.3
itsdangerous==2.2.0
blinker==1.8.2
click==8.1.7
